// apps/worker/src/types/workflowStep.ts
export enum WorkflowStep {
  START = "START",

  WAITING_SELECTION = "WAITING_SELECTION",

  WAITING_CUSTOMER_DETAILS = "WAITING_CUSTOMER_DETAILS",

  WAITING_ADDRESS = "WAITING_ADDRESS",

  WAITING_CONFIRMATION = "WAITING_CONFIRMATION",

  WAITING_PAYMENT = "WAITING_PAYMENT",

  WAITING_APPOINTMENT_SLOT = "WAITING_APPOINTMENT_SLOT",

  WAITING_RESERVATION_DETAILS = "WAITING_RESERVATION_DETAILS",

  COMPLETED = "COMPLETED",

  CANCELLED = "CANCELLED",

  
}